# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Hippo Equity Research API
  description: |
    Production-ready, scalable web application for global stock market data aggregation and analysis.
    
    The API provides endpoints for:
    - Stock data fetching and normalization
    - Multi-language content translation
    - Autopilot queue processing
    - PDF report generation
    - User preferences management
  version: 1.0.0
  contact:
    name: API Support
    email: support@hippo-equity.com
  license:
    name: Proprietary
    url: https://hippo-equity.com/license
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.hippo-equity.com
    description: Production server
...
paths:
...
  /api/queue/enqueue:
    post:
      tags:
        - Queue
      summary: Enqueue symbols for processing
      description: Adds stock symbols to the autopilot queue for batch processing
      operationId: enqueueSymbols
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnqueueRequest'
      responses:
        '200':
          description: Symbols enqueued successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnqueueResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/queue/status/{queueId}:
    get:
      tags:
        - Queue
      summary: Get queue status
      description: Returns the current status of a queue including progress and ETA
      operationId: getQueueStatus
      parameters:
        - name: queueId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Queue status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueStatus'
...
components:
  schemas:
...
    EnqueueRequest:
      type: object
      required:
        - symbols
      properties:
        symbols:
          type: array
          items:
            type: string
          minItems: 1
          example: [AAPL, GOOGL, MSFT]
    EnqueueResponse:
      type: object
      properties:
        queueId:
          type: string
          format: uuid
        symbols:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, processing, completed, failed]
        createdAt:
          type: string
          format: date-time
    QueueStatus:
      type: object
      properties:
        queueId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, processing, completed, failed]
        progress:
          type: number
          minimum: 0
          maximum: 100
        currentTask:
          type: string
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/QueueTask'
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
    QueueTask:
      type: object
      properties:
        id:
          type: string
        symbol:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed]
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: string
        timestamp:
          type: string
          format: date-time